test(library): :white_check_mark: Add sanity check.
A sanity check or sanity test is a basic test to quickly evaluate whether a claim or the result of a calculation can possibly be true. It is a simple check to see if the produced material is rational (that the material's creator was thinking rationally, applying sanity).  ― Wiki

Smoke testing.

Signed-off-by: Benjamin Vincent Kasapoglu <luxcium@users.noreply.github.com>