Add interfaces for JSON-RPC 2.0 Specification.
Copyright © 2023 Luxcium✨ (Benjamin Vincent)

derivative of:

Copyright (C) 2007-2010 by the JSON-RPC Working Group

This document and translations of it may be used to implement
JSON-RPC, it may be copied and furnished to others, and derivative
works that comment on or otherwise explain it or assist in its
implementation may be prepared, copied, published and distributed, in
whole or in part, without restriction of any kind, provided that the
above copyright notice and this paragraph are included on all such
copies and derivative works. However, this document itself may not
bemodified in any way.

The limited permissions granted above are perpetual and will not be
revoked.

This document and the information contained herein is provided "AS IS"
and ALL WARRANTIES, EXPRESS OR IMPLIED are DISCLAIMED, INCLUDING BUT
NOT LIMITED TO ANY WARRANTY THAT THE USE OF THE INFORMATION HEREIN
WILL NOT INFRINGE ANY RIGHTS OR ANY IMPLIED WARRANTIES OF
MERCHANTABILITY OR FITNESS FOR A PARTICULAR PURPOSE.

Signed-off-by: Luxcium (Benjamin Vincent) <luxcium@users.noreply.github.com>